# seminar-06/chat-client-java/src/main/openapi/openapi.yaml
---
openapi: 3.1.0
components:
  schemas:
    BackgroundColorEnum:
      type: string
      enum:
      - lightgray
      - white
      - aquamarine
      - lightyellow
      - lightblue
      - "#ffe4c4"
      description: enumeration of allowed background colors
      title: background color
    ChatMessage:
      type: object
      required:
      - id
      - timestamp
      - text
      description: represents a message in a chat
      title: chat message
      properties:
        id:
          type: string
          examples:
          - 123e4567-e89b-12d3-a456-426614174000
          pattern: \S
          description: uuid
        timestamp:
          $ref: "#/components/schemas/ZonedDateTime"
          type: string
          examples:
          - 2022-12-22T12:04:04.493908908+01:00
          pattern: \S
          description: time of creation
        text:
          type: string
          examples:
          - Hello! 😀
          pattern: \S
          description: text of message
        author:
          type: string
          examples:
          - John
          description: author of message
        textColor:
          type: string
          examples:
          - black
          description: HTML color name or RGB hex code
        backgroundColor:
          $ref: "#/components/schemas/BackgroundColorEnum"
    NewChatMessageRequest:
      type: object
      required:
      - text
      description: |
        Object for requesting new message.
        **Text** of the message must be located in the request body because URLs are limited in size.
        See [What is the maximum length of a URL in different browsers?](https://stackoverflow.com/questions/417142/what-is-the-maximum-length-of-a-url-in-different-browsers)
      properties:
        text:
          type: string
          examples:
          - Hello! 😀
          pattern: \S
          description: text of message
        textColor:
          type: string
          enum:
          - black
          - blue
          - darkgrey
          examples:
          - black
          description: HTML color name or RGB hex code
        backgroundColor:
          $ref: "#/components/schemas/BackgroundColorEnum"
    ZonedDateTime:
      type: string
      format: date-time
info:
  contact:
    email: jdoe@ics.muni.cz
    name: John Doe
    url: https://www.muni.cz/lide/666-joe-doe
  description: |
    Simple service for chatting. The API has operations for:
    * getting all messages
    * creating a new message
    * getting messages in pages, i.e. last 20 messages or previous 20 messages
    * getting a specific message by its id

    The description of the API in OpenAPI format is generated by
    the tool [SpringDoc v2](https://springdoc.org/) from annotated Java classes,
    mainly the controller class
    [ChatRestController](https://github.com/martin-kuba/muni-chat-service/blob/main/chat-server/src/main/java/cz/muni/chat/server/rest/ChatRestController.java)
    and the DTO classes in
    [cz.muni.chat.server.facade](https://github.com/martin-kuba/muni-chat-service/tree/main/chat-server/src/main/java/cz/muni/chat/server/facade)
    package.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Example Chat Service
  version: "1.1"
servers:
- url: "{scheme}://{server}:{port}"
  description: my server
  variables:
    scheme:
      enum:
      - http
      - https
      default: http
    server:
      default: localhost
    port:
      default: "8080"
paths:
  /api/message/{id}:
    get:
      summary: Returns identified message
      description: Looks up a message by its id.
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
        "404":
          description: message not found
          content: {}
      tags:
      - Chat Resource
  /api/messages:
    get:
      summary: Get all messages
      description: |
        Returns an array of objects representing chat messages, ordered from the newest to the oldest.
        Each message must have a **text** and **timestamp**, and optionally may have an **author**,
        a **text color** and a **background color**.
        It is possible to use [MarkDown](https://www.markdownguide.org/) in descriptions.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
      tags:
      - Chat Resource
    post:
      summary: Create a new message
      description: |
        Receives data both in request body and as URL parameter and stores them as a new message.
        Returns the new message as its response.
      parameters:
      - name: author
        in: query
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewChatMessageRequest"
        required: true
      responses:
        "201":
          description: New message was created successfully
        "400":
          description: input data were not correct
          content: {}
      tags:
      - Chat Resource
  /api/paged:
    get:
      summary: Paged messages
      description: |
        Returns a page of chat messages.
        The parameter `page` specifies zero-based index of the requested page,
        the parameter `size` specifies the size of the page.
        The parameter `sort` is ignored, sorting is always from the newest to the oldest.
      parameters:
      - name: pageOffset
        in: query
        schema:
          type: integer
          format: int32
      - name: pageSize
        in: query
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
      tags:
      - Chat Resource
  /chat:
    get:
      responses:
        "200":
          description: OK
          content:
            text/html: {}
      summary: Chat
      tags:
      - Chat UI Resource